Make emulator team owner of emu-dev
This makes the emulator team the owners of the emu-dev branch. The
emulator team needs to support libfdt on both MacOs and Windows MSVC,
which is not required for android main. Libfdt is a requirement for
building Qemu 8.

Having a separate branch will enable the emulator team to update FDT
independently of android main.

Bug: 297352758
Change-Id: I13fc0915b4e287c4f691a7366184e008fa237fdc